requirements.txt
psycopg2-binary
torch
torchvision
transformers
pillow
tqdm
scikit-learn
matplotlib
opencv-python
numpy
certifi
